+++ src/core/forms/FormField/FormField.tsx
@@ -1,7 +1,8 @@
 import React, { ReactNode, useEffect } from 'react';
-import { Checkbox, DatePicker, Form, Input, Radio, Switch, TimePicker,  Select as AntSelect } from 'antd';
+import { Button, Card, Checkbox, DatePicker, Form, Input, Radio, Switch, TimePicker,  Select as AntSelect } from 'antd';
 import { OptionSelector, IFieldOptions, IOptions } from './OptionSelector';
 import { useApi, useUi24Config } from '../../context';
+import { CloseOutlined } from '@ant-design/icons';
 import { CustomColorPicker } from '../../common/CustomColorPicker';
 import { IModalConfig } from '../../../modal/Modal';
 
@@ -11,6 +12,7 @@ export type IFormFieldType = "text" | "password" | "email" | "textarea" | "check
 
 
 interface IFormField {
+    namePrefixPath?: any[];
     name: string; //unique identifier, should be without spaces
     validationRules?: Array<any>; //rules matching ant design convention
     placeholder: string; //placeholder text
@@ -21,18 +23,47 @@ interface IFormField {
     label: string;
     style?: React.CSSProperties;
     initialValue?: any;
-    setFormValue?: Function
+    setFormValue?: Function;
+
+    // for list and map fields
+    type?: string;
+    properties?: Array<IFormField>
+    items?: {
+        type: string,
+        properties?: Array<IFormField>
+    }
 }
 
 const { TextArea } = Input;
-export function FormField( {fieldType = "text", name, validationRules, label = "", prefixIcon, placeholder = "", options = [], style, initialValue, setFormValue, addNewOption, ...restFormItemProps } : IFormField ) {
-    
-    const { selectConfig } = useUi24Config()
-    const formatConfig = selectConfig( config => config.formatConfig )
 
-    return <div style={{ marginBottom: "24px" }} key={"CustomFormFields"}>
-      <Form.Item name={ name } rules={ validationRules } label={ label } style={ style } initialValue={initialValue} >
-        { fieldType === "text" && <Input type={ fieldType || "text" } prefix={ prefixIcon } placeholder={ placeholder } />}
+const makeFormItem = ({
+        fieldType = "text", 
+        namePrefixPath,
+        name, 
+        validationRules, 
+        label = "", 
+        prefixIcon, 
+        placeholder = "", 
+        options = [], 
+        style, 
+        initialValue, 
+        setFormValue,
+        addNewOption,
+        ...restFormItemProps
+    }: IFormField) => {
+    
+    const { selectConfig } = useUi24Config();
+    const formatConfig = selectConfig( config => config.formatConfig );
+    return <>
+        <Form.Item 
+            name={ namePrefixPath?.length ? [...namePrefixPath, name] : name } 
+            rules={ validationRules } 
+            label={ label } 
+            style={ style } 
+            initialValue={initialValue} 
+        >
+
+        { fieldType === "text" && <Input type={ fieldType || "text" } prefix={ prefixIcon } placeholder={ placeholder } /> }
         { fieldType === "textarea" && <TextArea placeholder={ placeholder } />}
         { fieldType === "password" && <Input.Password type={ fieldType || "password" } prefix={ prefixIcon } placeholder={ placeholder } /> }
         { fieldType === "email" && <Input type={ fieldType || "email" } prefix={ prefixIcon } placeholder={ placeholder } /> }
@@ -64,7 +95,7 @@ export function FormField( {fieldType = "text", name, validationRules, label = "
                 listType={ restFormItemProps['listType'] ?? 'picture-card'} 
                 withImageCrop = {restFormItemProps['withImageCrop'] ?? true} 
 
-                 // config for the default image uploader
+                // config for the default image uploader
                 fileNamePrefix = { restFormItemProps['fileNamePrefix'] ?? undefined}
                 getSignedUploadUrlAPIConfig  = { restFormItemProps['getSignedUploadUrlAPIConfig'] ?? undefined}
             />
@@ -87,8 +118,62 @@ export function FormField( {fieldType = "text", name, validationRules, label = "
                 uploadFile = { restFormItemProps['uploadFile'] ?? undefined}
             />
         }
+    </Form.Item>
+    </>
+}
 
-      </Form.Item>
+const makeFormListItem = ({
+    name, 
+    namePrefixPath,
+    validationRules, 
+    label = "", 
+    initialValue, 
+    items,
+}: IFormField) => {
+    return <>
+        <Form.List 
+        name={ namePrefixPath?.length ? [...namePrefixPath, name] : name } 
+            rules={ validationRules } 
+            initialValue={initialValue} 
+        >
+        {(fields, { add, remove }) => (
+            <div style={{ display: 'flex', rowGap: 16, flexDirection: 'column' }}>
+                {fields.map((field) => (
+                    <Card
+                        size="small"
+                        title={`${label} ${field.name + 1}`}
+                        key={field.key}
+                        extra={ <CloseOutlined onClick={() => { remove(field.name); }} /> }
+                    >
+                        {
+                            items.properties.map( (property) => {
+                                return makeFormItem({
+                                    ...property,
+                                    namePrefixPath: namePrefixPath?.length ? [...namePrefixPath, field.name] : [field.name]
+                                })
+                            })
+                        }
+                    </Card>
+                ))}
+
+                <Button type="dashed" onClick={() => add()} block> + Add {label} </Button>
+            </div>
+        )}
+        </Form.List>
+    </>
+}
+export function FormField( formField : IFormField ) {
+
+    const {
+        fieldType = "text",
+        type,
+    } = formField;
+
+    return <div style={{ marginBottom: "24px" }} key={"CustomFormFields"}>
+        { ( type === 'list' && !['wysiwyg', 'rich-text'].includes(fieldType.toLocaleLowerCase()) ) 
+            ? makeFormListItem(formField) 
+            : makeFormItem(formField) 
+        }
     </div>
 }
 
@@ -109,6 +194,13 @@ interface IFormFieldResponse {
     getSignedUploadUrlAPIConfig ?: GetSignedUploadUrlAPIConfig,
     withImageCrop?: boolean;
 
+    // list and map fields
+    type?: string;
+    properties?: Array<IFormFieldResponse>
+    items?: {
+        type: string,
+        properties?: Array<IFormFieldResponse>
+    }
 }
 
 const convertValidationRules = ( validationRules : Array<IPreDefinedValidations> ) => {
@@ -150,7 +242,14 @@ export const convertColumnsConfigForFormField  = ( columnsConfig : Array<IFormFi
             withImageCrop: columnConfig.withImageCrop,
             fileNamePrefix:  columnConfig.fileNamePrefix,
             getSignedUploadUrlAPIConfig: columnConfig.getSignedUploadUrlAPIConfig,
-          
+
+            // for list and map fields
+            type: columnConfig.type,
+            properties: columnConfig.properties ? convertColumnsConfigForFormField(columnConfig.properties) : [],
+            items: {
+                ...(columnConfig.items || {}),
+                properties: columnConfig.items?.properties ? convertColumnsConfigForFormField(columnConfig.items.properties) : []
+            }
       } as IFormField
   })
 }