+++ src/pages/Auth/LoginPage.tsx
@@ -19,6 +19,10 @@ const LoginForm = () => {
   const { propertiesConfig, apiConfig } = usePageConfig("/login");
   const [ loader, setLoader ] = useState<boolean>( false )
   const { callApiMethod } = useApi();
+  const {propertiesConfig:signupPropertiesConfig} = usePageConfig("/signup");
+  const {propertiesConfig:verifyPropertiesConfig} = usePageConfig("/verify");
+  const {propertiesConfig:forgotPasswordPropertiesConfig} = usePageConfig("/forgot-password");
+
 
   const onFinish = async (payload: any) => {
     setLoader(true)
@@ -51,29 +55,34 @@ const LoginForm = () => {
     disabled={ loader }
     buttonLoader={ loader }
   >
+      
+    { forgotPasswordPropertiesConfig?.length &&
       <div className="PreAuthLoginActions">
           <Checkbox onChange={onChange}>Remember Me</Checkbox>
           <Link className="forgotPassword" title="Forgot Password ?" url='/forgot-password' />
       </div>
+    }
 
-      <div className="PreAuthLoginActions">
+    <div className="PreAuthLoginActions">
+      {signupPropertiesConfig?.length &&
         <Button 
-              type = "default"
-              size = "middle"
-              style = {{ width: "48%", margin:"1%" }}
-              onClick = {handleRegister}
-          > 
-            Create Account 
-          </Button>
-
-          <Button 
-              type = "dashed"
-              size = "middle"
-              style = {{ width: "48%", margin:"1%" }}
-              onClick = {handleVerification}
-          > 
-            Verify Account 
-          </Button>
-      </div>
+            type = "default"
+            size = "middle"
+            style = {{ width: "48%", margin:"1%" }}
+            onClick = {handleRegister}
+        > 
+          Create Account 
+        </Button>
+      }
+      { verifyPropertiesConfig?.length &&  <Button 
+            type = "dashed"
+            size = "middle"
+            style = {{ width: "48%", margin:"1%" }}
+            onClick = {handleVerification}
+        > 
+          Verify Account 
+        </Button>
+      }
+    </div>
   </AuthForm>
 }